+++ src/Service/Provider/MicrosoftEntraIdOidcClientEntityWriter.php
@@ -6,14 +6,11 @@ namespace Heptacom\AdminOpenAuth\Service\Provider;
 
 use Heptacom\AdminOpenAuth\Component\Provider\DefaultInjectingEntityWriterDecorator;
 
-/**
- * @deprecated tag:v8.0.0 - Will be renamed into MicrosoftEntraIdOidcClientEntityWriter
- */
-final class MicrosoftAzureOidcClientEntityWriter extends DefaultInjectingEntityWriterDecorator
+final class MicrosoftEntraIdOidcClientEntityWriter extends DefaultInjectingEntityWriterDecorator
 {
     protected function getProvider(): string
     {
-        return MicrosoftAzureOidcProvider::PROVIDER_NAME;
+        return MicrosoftEntraIdOidcProvider::PROVIDER_NAME;
     }
 
     protected function injectDefaults(array $payload): array