+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Refactored role assignment rules to use the new login actions
 * Replaced deprecated JWT libraries (`web-token/jwt-core`, `web-token/jwt-signature*`) with replacement library (`web-token/jwt-library`) (GitHub Issue #29)
 * Changed user creation to create new users without admin privileges. The privileges are applied later in the login process. (GitHub Issue #12)
+* Technically renamed Azure AD to Entra ID (see deprecations from [v7.0.0](#700))
 
 # 7.0.1